Moving check run logs output into it's own component
