+++ app/src/ui/check-runs/ci-check-list-item.tsx
@@ -1,23 +1,9 @@
 import * as React from 'react'
-import {
-  getCheckDurationInSeconds,
-  IRefCheck,
-  IRefCheckOutput,
-  RefCheckOutputType,
-} from '../../lib/stores/commit-status-store'
-
+import { IRefCheck } from '../../lib/stores/commit-status-store'
 import { Octicon } from '../octicons'
 import { getClassNameForCheck, getSymbolForCheck } from '../branches/ci-status'
 import classNames from 'classnames'
-import { APICheckConclusion } from '../../lib/api'
-import { Button } from '../lib/button'
-import { encodePathAsUrl } from '../../lib/path'
-
-// TODO: Get empty graphic for logs?
-const BlankSlateImage = encodePathAsUrl(
-  __dirname,
-  'static/empty-no-pull-requests.svg'
-)
+import { CICheckRunLogs } from './ci-check-run-item-logs'
 
 interface ICICheckRunListItemProps {
   /** The check run to display **/
@@ -51,151 +37,8 @@ export class CICheckRunListItem extends React.PureComponent<
     this.props.onViewOnGitHub(this.props.checkRun)
   }
 
-  private isNoAdditionalInfoToDisplay(output: IRefCheckOutput): boolean {
-    return (
-      this.isNoOutputText(output) &&
-      (output.summary === undefined ||
-        output.summary === null ||
-        output.summary.trim() === '')
-    )
-  }
-
-  private isNoOutputText(output: IRefCheckOutput): boolean {
-    return (
-      output.type === RefCheckOutputType.Default &&
-      (output.text === null || output.text.trim() === '')
-    )
-  }
-
-  private renderActionsLogOutput = (output: IRefCheckOutput) => {
-    if (output.type === RefCheckOutputType.Default) {
-      return null
-    }
-
-    return output.steps.map((step, i) => {
-      const header = (
-        <div className="ci-check-run-log-step" key={i}>
-          <div className="ci-check-status-symbol">
-            <Octicon
-              className={classNames(
-                'ci-status',
-                `ci-status-${getClassNameForCheck(step)}`
-              )}
-              symbol={getSymbolForCheck(step)}
-              title={step.name}
-            />
-          </div>
-          <div className="ci-check-run-log-step-name">{step.name}</div>
-          <div>{getCheckDurationInSeconds(step)}s</div>
-        </div>
-      )
-
-      const log =
-        step.conclusion === APICheckConclusion.Failure ? step.log : null
-
-      return (
-        <>
-          {header}
-          {log}
-        </>
-      )
-    })
-  }
-
-  private renderNonActionsLogOutput = (output: IRefCheckOutput) => {
-    if (output.type === RefCheckOutputType.Actions || output.text === null) {
-      return null
-    }
-
-    // TODO: Html needs santized. Later PR
-    return <div dangerouslySetInnerHTML={{ __html: output.text }}></div>
-  }
-
-  private renderMetaOutput = (
-    output: IRefCheckOutput,
-    checkRunName: string
-  ) => {
-    const { title, summary } = output
-
-    // Don't display something empty or redundant
-    const displayTitle =
-      title !== null &&
-      title.trim() !== '' &&
-      title.trim().toLocaleLowerCase() !==
-        checkRunName.trim().toLocaleLowerCase()
-
-    const displaySummary =
-      summary !== null && summary !== undefined && summary.trim() !== ''
-
-    return (
-      <div>
-        {displayTitle ? <div>{title}</div> : null}
-        {displaySummary ? <pre>{summary}</pre> : null}
-      </div>
-    )
-  }
-
-  private renderEmptyLogOutput = () => {
-    return (
-      <div className="no-logs-to-display">
-        No additional information to display.
-      </div>
-    )
-  }
-
-  private renderLoadingLogs = () => {
-    return (
-      <div className="loading-logs">
-        <img src={BlankSlateImage} className="blankslate-image" />
-        <div className="title">Hang tight</div>
-        <div className="loading-blurb">Loading the logs as fast as I can!</div>
-      </div>
-    )
-  }
-
-  private renderViewOnGitHub = () => {
-    return (
-      <div className="view-on-github">
-        <Button onClick={this.onViewOnGitHub}>View on GitHub</Button>
-      </div>
-    )
-  }
-
-  private hasActionsWorkflowLogs() {
-    return this.props.checkRun.actionsWorkflowRunId !== undefined
-  }
-
-  private renderLogs = () => {
-    const {
-      loadingActionLogs,
-      loadingActionWorkflows,
-      checkRun: { output, name },
-    } = this.props
-
-    if (
-      loadingActionWorkflows ||
-      (this.hasActionsWorkflowLogs() && loadingActionLogs)
-    ) {
-      return this.renderLoadingLogs()
-    }
-
-    return (
-      <div className="ci-check-list-item-logs">
-        <div className="ci-check-list-item-logs-output">
-          {this.isNoAdditionalInfoToDisplay(output)
-            ? this.renderEmptyLogOutput()
-            : null}
-          {this.renderMetaOutput(output, name)}
-          {this.renderActionsLogOutput(output)}
-          {this.renderNonActionsLogOutput(output)}
-        </div>
-        {this.renderViewOnGitHub()}
-      </div>
-    )
-  }
-
   public render() {
-    const { checkRun, showLogs } = this.props
+    const { checkRun, showLogs, loadingActionLogs } = this.props
 
     return (
       <>
@@ -221,7 +64,14 @@ export class CICheckRunListItem extends React.PureComponent<
             </div>
           </div>
         </div>
-        {showLogs ? this.renderLogs() : null}
+        {showLogs ? (
+          <CICheckRunLogs
+            checkRun={checkRun}
+            loadingActionLogs={loadingActionLogs}
+            loadingActionWorkflows={loadingActionLogs}
+            onViewOnGitHub={this.onViewOnGitHub}
+          />
+        ) : null}
       </>
     )
   }
